Spelling: The member, non plural, of circle.
This was initially a heads-up from javier.llamas on the translation platform about line 64.

"Successfully" also removed. If there is a non-successful way of doing things, or if it failed beforehand, i can reinstate it.

Being expelled/excluded from a circle rather than removed makes more sense for me, but i opted out of it since it is just more advanced for not a whole lot of reward, and potential problems in translation.