honor user enumeration prevention of the share api
If user enumeration is disabled in nextcloud core, then circle can
be used to circumvent this measure. For example the adduser button
in the user facing circle UI allows enumeration of all registered
users.

This patch honors the choice made in
 'shareapi_allow_share_dialog_user_enumeration'
for deciding if auto completion should present partial results.
This is in line with other apps, such as webdav, which reuse this
configuration choice to disable user enumeration.

In case this preference is set, all partial results are removed
from the results.

Fixes #152