fix(ShareWrapperRequest): Save share attributes on creation
- Before this commit, attributes added to a share were ignored when the share type was team/circles (share type `7`).

- Share attributes are not expected to contain any key or property named `enabled`.

Partial resolution of: https://github.com/nextcloud/server/issues/46920

This is a partial resolution because setting the `download` attribute may not fully enforce the correct download permissions. A follow-up PR in the server may be needed to update the `hideDownload` property of the share based on the value of `download` in the share attributes.

Signed-off-by: fenn-cs <fenn25.fn@gmail.com>