fix(CoreQueryBuilder): Use correct member entry for circle as initiator
Before, the member entry that matched the `singleId` of the user and
`level = 9` was used as initiator, which in practice means the member
entry for the internal user circle of the user most of the time.

Instead, we want to use the member entry that matches `singleId` of
the user and `circleId` of the circle in question.

This fixes the wrong `level` being set for `initiator` when calling
`circleProbe()` with `DataProbe::INITIATOR`.

Fixes: #1757

Signed-off-by: Jonas <jonas@freesources.org>