fix: Set a token to shares retreived from the DB
When accessing a circle share from the public share view, the `addOpenGraphHeaders` method was erroring in `DefaultPublicShareTemplateProvider`.
This is due to the fact that the `$share` object for circles share does not contain a token, so the computation of the `$shareUrl` was failing.

This commit ensures that the share has a token.

Signed-off-by: Louis Chemineau <louis@chmn.me>